add  README file and 배포했을때 file

+++ routers/post.js
@@ -31,11 +31,19 @@ router.post('/add/content', async (req, res) => {
   }
 });
 // 게시글 삭제하기
-router.delete('/delete', async (req, res) => {
+router.delete('/delete/:postID', async (req, res) => {
   try {
-    const { postID } = req.body;
+    const { postID } = req.params;
+    const { postPassword } = req.body;
+
+    const detailPost = await Post.find({ postID: postID }, { _id: false });
+    // DB에 저장된 비밀번호와 일치하지 않은 경우
+    if (detailPost[0].postPassword != postPassword) {
+      res.send({ msg: '비밀번호가 일치하지 않습니다.', result: false });
+      return;
+    }
     await Post.deleteOne({ postID: postID });
-    res.send({ msg: '해당 게시글의 삭제를 완료하였습니다.' });
+    res.send({ msg: '해당 게시글의 삭제를 완료하였습니다.', result: true });
   } catch (error) {
     console.log(error);
   }
@@ -46,17 +54,19 @@ router.put('/modify/:postID', async (req, res) => {
   try {
     const { postID } = req.params;
     const { postTitle, postContent, postPassword } = req.body;
+    const detailPost = await Post.find({ postID: postID }, { _id: false });
 
-    // 비밀번호가 일치하지 않을 경우
-    if (await !isMatchPwd()) {
-      res.send({ msg: '비밀번호가 일치하지 않습니다.' });
+    // DB에 저장된 비밀번호와 일치하지 않은 경우
+    if (detailPost[0].postPassword != postPassword) {
+      res.send({ msg: '비밀번호가 일치하지 않습니다.', result: false });
       return;
     }
+
     await Post.updateOne(
       { postID: postID },
       { $set: { postTitle: postTitle, postContent: postContent } }
     );
-    res.send({ msg: '해당 게시글의 수정을 완료하였습니다.' });
+    res.send({ msg: '해당 게시글의 수정을 완료하였습니다.', result: true });
   } catch (error) {
     console.log(error);
   }
@@ -70,13 +80,6 @@ function getCurrentTime() {
 
   const currentTime = `${year}.${month}.${date}`;
   return currentTime;
-}
-
-//비밀번호 DB에 있는 것과 일치하는지 확인
-function isMatchPwd(password){
-  return new Promise((resolve, reject) => {
-      
-  })
 }
 
 module.exports = router;
